+++ src/App.tsx
@@ -9,9 +9,48 @@ import { SearchPage } from './pages/SearchPage';
 import { NavBar } from './components/NavBar';
 
 const App = () => {
+  // state variables
   const [businessData, setBusinessData, loadingBusinesses] =
     useData<BusinessEntry[]>('/');
   const [filteredData, setFilteredData] = useState<BusinessEntry[]>([]);
+  const [searchText, setSearchText] = useState<string>('');
+  const [advancedFilterValues, setAdvancedFilterValues] = useState<string[]>(
+    [] as string[]
+  );
+
+  // filtering businesses
+  const advancedFilteredBusinesses = Object.values(businessData).filter(
+    (business) => {
+      for (const value of advancedFilterValues) {
+        if (
+          business.hasOwnProperty('Initiatives') &&
+          business.Initiatives.includes(value)
+        ) {
+          return true;
+        }
+      }
+      return false;
+    }
+  );
+
+  const filteredText = searchText.split(' ');
+
+  const intersect = (keywords: Array<string>, tags: Array<string>) =>
+    keywords.filter((keyword) => tags.some((tag) => tag.includes(keyword)));
+
+  const finalFilteredBusinesses = Object.values(
+    advancedFilterValues.length > 0 ? advancedFilteredBusinesses : businessList
+  ).filter(
+    (business) =>
+      intersect(
+        filteredText.map((text) => text.toLowerCase()),
+        business['Search Tags']
+          .concat([business.Title, business.Description])
+          .map((text) => text.toLowerCase())
+      ).length > 0
+  );
+
+  // app page selection
   if (loadingBusinesses || !businessData)
     return (
       <div>
